clean up common.h
