+++ oi/ContainerTypeEnum.h
@@ -14,16 +14,6 @@
  * limitations under the License.
  */
 #pragma once
-#include <map>
-#include <set>
-#include <string>
-#include <vector>
-
-extern "C" {
-#include <drgn.h>
-}
-
-constexpr int oidMagicId = 0x01DE8;
 
 #define LIST_OF_CONTAINER_TYPES  \
   X(UNKNOWN_TYPE)                \
@@ -73,46 +63,3 @@ enum ContainerTypeEnum {
   LIST_OF_CONTAINER_TYPES
 #undef X
 };
-
-struct RootInfo {
-  std::string varName;
-  struct drgn_qualified_type type;
-};
-
-struct ClassMember {
-  std::string typeName;
-  std::string varName;
-};
-
-struct DrgnClassMemberInfo {
-  struct drgn_type* type;
-  std::string member_name;
-  uint64_t bit_offset;
-  uint64_t bit_field_size;
-  bool isStubbed;
-};
-
-struct TypeHierarchy {
-  std::map<struct drgn_type*, std::vector<DrgnClassMemberInfo>> classMembersMap;
-  std::map<
-      struct drgn_type*,
-      std::pair<ContainerTypeEnum, std::vector<struct drgn_qualified_type>>>
-      containerTypeMap;
-  std::map<struct drgn_type*, struct drgn_type*> typedefMap;
-  std::map<std::string, size_t> sizeMap;
-  std::set<struct drgn_type*> knownDummyTypeList;
-  std::map<struct drgn_type*, struct drgn_type*> pointerToTypeMap;
-  std::set<struct drgn_type*> thriftIssetStructTypes;
-  std::map<struct drgn_type*, std::vector<struct drgn_type*>> descendantClasses;
-};
-
-// Helper for std::variant and std::visit
-// https://en.cppreference.com/w/cpp/utility/variant/visit
-template <class... Ts>
-struct visitor : Ts... {
-  using Ts::operator()...;
-};
-
-// Type deduction for the helper above
-template <class... Ts>
-visitor(Ts...) -> visitor<Ts...>;
